move shared styled components to their own file
